requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
openai==1.35.10
requests==2.32.3
beautifulsoup4==4.12.3
httpx==0.27.2
